image-builder: pin notary signers to specific job types (#6917)
This PR adds ability to enable specific signer service only for specific ProwJob type.
When the signer configuration contains a job-type field, it is considered as CI-based service. Its invocation is only restricted to CI mode for a specific repository, or globally.
* If signer contains a job-type field, and is running in Prow, image-builder will determine if this service should sign an image based on the value of job-type and env variable `JOB_TYPE`
* If signer contains job-type field, but it's not running on CI, image-builder ignores signing using this service
* If signer doesn't contain a job-type field it will be used and enabled by default

This approach allows to pin specific notary service to each of job-types, eg. prod service only for postsubmit jobs, and dev service for all types of job

/area tools
/kind feature