Tekton golangci-lint task (#7278)
* feat(golangci-lint): add golangci-lint task
This commit adds a new task for golangci-lint. The task runs a golang linter and is designed to be used in Tekton pipelines. The task is versioned at 0.1 and includes a description, workspaces, and parameters. The task uses the golangci-lint:v20230221-85b2d35f image and has resource limits of 1Gi memory and 1 CPU.

* Changed limits to requests and increased values to avoid timeouts.

* feat(golangci-lint): add golanci-lint Task. The golangci-lint Task is a Tekton task that runs a golang linter. It is designed to be triggered by prow as a prowjob. The task uses the golangci-lint tool.

Improve task description and workspace description.

The task description has been updated to include information about the prowjob trigger and the use of the golangci-lint tool.
The workspace description has been updated to include information about the default mount path and the fact that it should be a shared workspace.
Information about the golangci-lint tool and its configuration has been added to the task description.

Add sample prowjob for linting go code.

* Fixed typos.

* Apply suggestions from code review

Co-authored-by: Iwona Langer <iwona.langer@sap.com>

* Updated description with more details about tool behaviour.

* Added more details about task capabilities.

* Apply suggestions from code review

Co-authored-by: Iwona Langer <iwona.langer@sap.com>

* Fixed tool name in description.
Removed timeout as it can't be set by parameters. A default value will be aligned with prow kubernetes agent.

---------

Co-authored-by: Iwona Langer <iwona.langer@sap.com>